moving doctrine cache plugin to separate file
