+++ lib/Xframe/Core/DependencyInjectionContainer.php
@@ -2,6 +2,7 @@
 
 namespace Xframe\Core;
 
+use Doctrine\Common\Cache\Cache;
 use Doctrine\Common\EventManager;
 use Doctrine\ORM\EntityManager;
 use Memcache;
@@ -22,6 +23,7 @@ use Xframe\Request\FrontController;
  * @property PDO $database
  * @property Container $plugin
  * @property Memcache $cache
+ * @property Cache $doctrineCache
  * @property EntityManager $em
  * @property EventManager $evm
  *
